requirements.txt
allure-pytest==2.8.17
allure-python-commons==2.8.17
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
blinker==1.4
Brotli==1.0.9
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
colorama==0.4.4
cryptography==3.2.1
cssselect==1.1.0
cssutils==2.3.0
Flask==1.1.4
h11==0.12.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.8.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpath==0.82
kaitaistruct==0.9
ldap3==2.8.1
loguru==0.5.1
lxml==4.6.4
MarkupSafe==2.0.1
more-itertools==8.10.0
msgpack==1.0.2
packaging==21.2
passlib==1.7.4
pluggy==0.13.1
premailer==3.10.0
protobuf==3.13.0
publicsuffix2==2.20191221
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydivert==2.1.0
PyMySQL==1.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.2
pytest==6.0.1
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
six==1.16.0
sortedcontainers==2.2.2
toml==0.10.2
tornado==6.1
typing-extensions==3.10.0.2
urllib3==1.26.7
urwid==2.1.2
Werkzeug==1.0.1
win32-setctime==1.0.3
wsproto==0.15.0
xlrd==2.0.1
xlwt==1.3.0
yagmail==0.14.260
zipp==3.6.0
zstandard==0.14.1
